+++ NEWS.md
@@ -17,6 +17,7 @@
 - Matching case `case?` statements are supported in VHDL-2008 mode.
 - Fixed several bugs in the implementation of guarded signals.
 - Implemented VHDL-2008 rules for generic visibility.
+- Shared variable declaration permitted in entity declaration.
 
 ## Version 1.6.2 - 2022-04-03
 - Fix `make -j` with GNU make (#440).